removed all errors regarding indexing with non-integer values

+++ deltasigma/_ds_synNTFobj1.py
@@ -37,7 +37,7 @@ def ds_synNTFobj1(x, p, osr, f0):
     z = np.exp(2j*np.pi*(f0 + 0.5/osr*x))
     z = carray(z)
     if f0 > 0:
-        z = padt(z, p.shape[0]/2., np.exp(2j*np.pi*f0))
+        z = padt(z, int(p.shape[0]/2.), np.exp(2j*np.pi*f0))
 
     z = np.hstack((z, np.conj(z))) 
     z = z[:]
